PLAT-245: I removed all the redundant func tests that were testing Platform 1.0 behaviour
