+++ ledger-core/v2/application/functest/common_test.go
@@ -31,15 +31,15 @@ func TestWrongUrl(t *testing.T) {
 	jsonValue, _ := json.Marshal(postParams{})
 	testURL := launchnet.AdminHostPort + "/not_api"
 	postResp, err := http.Post(testURL, "application/json", bytes.NewBuffer(jsonValue))
-	defer postResp.Body.Close()
 	require.NoError(t, err)
+	defer postResp.Body.Close()
 	require.Equal(t, http.StatusNotFound, postResp.StatusCode)
 }
 
 func TestWrongJson(t *testing.T) {
 	postResp, err := http.Post(launchnet.TestRPCUrl, "application/json", bytes.NewBuffer([]byte("some not json value")))
-	defer postResp.Body.Close()
 	require.NoError(t, err)
+	defer postResp.Body.Close()
 	require.Equal(t, http.StatusOK, postResp.StatusCode)
 	body, err := ioutil.ReadAll(postResp.Body)
 	require.NoError(t, err)
