+++ ledger-core/v2/application/functest/get_status_test.go
@@ -25,3 +25,15 @@ func TestGetStatus(t *testing.T) {
 	require.Equal(t, "CompleteNetworkState", status.NetworkState)
 	require.Equal(t, numNodes, status.WorkingListSize)
 }
+
+func getStatus(t testing.TB) statusResponse {
+	body := getRPSResponseBody(t, launchnet.TestRPCUrl, postParams{
+		"jsonrpc": "2.0",
+		"method":  "node.getStatus",
+		"id":      "1",
+	})
+	rpcStatusResponse := &rpcStatusResponse{}
+	unmarshalRPCResponse(t, body, rpcStatusResponse)
+	require.NotNil(t, rpcStatusResponse.Result)
+	return rpcStatusResponse.Result
+}
