+++ ledger-core/v2/application/functest/main_func_test.go
@@ -8,22 +8,14 @@
 package functest
 
 import (
-	"fmt"
 	"os"
 	"testing"
-
-	"github.com/pkg/errors"
 
 	"github.com/insolar/assured-ledger/ledger-core/v2/application/testutils/launchnet"
 )
 
 func TestMain(m *testing.M) {
 	os.Exit(launchnet.Run(func() int {
-		err := setMigrationDaemonsRef()
-		if err != nil {
-			fmt.Println(errors.Wrap(err, "[ setup ] get reference daemons by public key failed ").Error())
-		}
-
 		return m.Run()
 	}))
 }
