--- ledger-core/v2/application/functest/member_create_test.go
@@ -1,48 +1,0 @@
-// Copyright 2020 Insolar Network Ltd.
-// All rights reserved.
-// This material is licensed under the Insolar License version 1.0,
-// available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
-
-// +build functest
-
-package functest
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/insolar/assured-ledger/ledger-core/v2/application/testutils/launchnet"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestMemberCreate(t *testing.T) {
-	member, err := newUserWithKeys()
-	require.NoError(t, err)
-	result, err := signedRequest(t, launchnet.TestRPCUrlPublic, member, "member.create", nil)
-	require.NoError(t, err)
-	output, ok := result.(map[string]interface{})
-	require.True(t, ok)
-	require.NotEqual(t, "", output["reference"])
-}
-
-func TestMemberCreateWithBadKey(t *testing.T) {
-	member, err := newUserWithKeys()
-	require.NoError(t, err)
-	member.PubKey = "fake"
-	_, err = signedRequestWithEmptyRequestRef(t, launchnet.TestRPCUrlPublic, member, "member.create", nil)
-	data := checkConvertRequesterError(t, err).Data
-	require.Contains(t, data.Trace, fmt.Sprintf("problems with decoding. Key - %s", member.PubKey))
-}
-
-func TestMemberCreateWithSamePublicKey(t *testing.T) {
-	member, err := newUserWithKeys()
-	require.NoError(t, err)
-
-	_, err = signedRequest(t, launchnet.TestRPCUrlPublic, member, "member.create", nil)
-	require.NoError(t, err)
-
-	_, err = signedRequestWithEmptyRequestRef(t, launchnet.TestRPCUrlPublic, member, "member.create", nil)
-	data := checkConvertRequesterError(t, err).Data
-	require.Contains(t, data.Trace, "can't set reference because this key already exists")
-}
